move out sidebar component
